Reorganised UI to new Katsu response
