+++ src/js/components/Search/Search.js
@@ -1,55 +1,52 @@
-import React, { useEffect } from 'react';
-import { Container, Row, Col } from 'react-bootstrap';
-import { Button } from 'antd';
+import React from 'react';
+import { Row, Col, Button, Typography } from 'antd';
 import { useSelector, useDispatch } from 'react-redux';
-import { addQueryableFields, makeGetKatsuPublic } from '../../features/query';
+import { makeGetKatsuPublic } from '../../features/search/query';
 import SearchFieldsStack from './SearchFieldsStack';
 import SearchResults from './SearchResults';
 
 const Search = () => {
   const dispatch = useDispatch();
 
+  const searchSections = useSelector((state) => state.query.querySections);
+  const buttonDisabled = useSelector((state) => state.query.queryParamCount) === 0;
+  const isFetchingData = useSelector((state) => state.query.isFetchingData);
+
   const queryKatsuPublic = () => {
     dispatch(makeGetKatsuPublic());
   };
 
-  const buttonDisabled =
-    useSelector((state) => state.query.queryParamCount) === 0;
-
-  const isFetchingData = useSelector((state) => state.query.isFetchingData);
-
   return (
-    <Container>
-      <Row>
-        <Col sm={{ span: 10 }}>
-          <SearchFieldsStack />
-        </Col>
-      </Row>
-      <Row>
-        <Col
-          className="text-center"
-          xs={{ span: 4, offset: 4 }}
-          md={{ span: 6, offset: 3 }}
-        >
-          <Button
-            type="primary"
-            onClick={queryKatsuPublic}
-            size="large"
-            shape="round"
-            loading={isFetchingData}
-            disabled={buttonDisabled}
-            style={{ margin: '10px' }}
-          >
-            Get Data
-          </Button>
-        </Col>
-      </Row>
-      <Row>
-        <Col className="text-center">
-          <SearchResults />
-        </Col>
-      </Row>
-    </Container>
+    <Row justify="center">
+      <Col style={{ width: '900px' }}>
+        {searchSections.map((e, i) => (
+          <div key={i} style={{ marginBottom: '20px' }}>
+            <Typography.Title level={4}>{e.section_title}</Typography.Title>
+            <SearchFieldsStack key={i} queryFields={e.fields} />
+          </div>
+        ))}
+        <Row justify="center">
+          <Col className="text-center">
+            <Button
+              type="primary"
+              onClick={queryKatsuPublic}
+              size="large"
+              shape="round"
+              loading={isFetchingData}
+              disabled={buttonDisabled}
+              style={{ margin: '20px' }}
+            >
+              Get Data
+            </Button>
+          </Col>
+        </Row>
+        <Row justify="center">
+          <Col>
+            <SearchResults />
+          </Col>
+        </Row>
+      </Col>
+    </Row>
   );
 };
