+++ src/js/components/Search/SelectOption.js
@@ -1,21 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import { Select } from 'antd';
 import { useDispatch } from 'react-redux';
-import { addQueryParam } from '../../../features/query';
-import { queryTypes } from '../../../constants/queryConstants';
+import { addQueryParam } from '../../features/search/query';
 
-const SelectOption = ({ name, isChecked, data }) => {
+const SelectOption = ({ id, isChecked, options }) => {
   const dispatch = useDispatch();
 
-  const [value, setValue] = useState('');
+  const [value, setValue] = useState(options[0]);
 
   useEffect(() => {
     if (isChecked) {
-      dispatch(
-        addQueryParam({ name, queryType: queryTypes.SELECT, params: { value } })
-      );
+      dispatch(addQueryParam({ id, value }));
     }
-  }, [isChecked, value, data]);
+  }, [isChecked, value, options]);
 
   const handleValueChange = (newValue) => {
     setValue(newValue);
@@ -23,15 +20,15 @@ const SelectOption = ({ name, isChecked, data }) => {
 
   return (
     <Select
-      id={name}
+      id={id}
       disabled={!isChecked}
       showSearch
       style={{ width: '100%' }}
       onChange={handleValueChange}
+      defaultValue={options[0]}
     >
-      <Select.Option key={name} value=""></Select.Option>
-      {data.enum.map((item) => (
-        <Select.Option key={item} value={item.name}>
+      {options.map((item) => (
+        <Select.Option key={item} value={item}>
           {item}
         </Select.Option>
       ))}
