--- src/js/constants/queryConstants.js
@@ -1,9 +1,0 @@
-export const queryTypes = {
-  DATE: 'date',
-  SELECT: 'select',
-  INPUT: 'input',
-  RANGE: 'range',
-  NUMBER_INPUT: 'numberInput',
-};
-
-export const QUERY_START_LIMIT = '2020-03-01';
